refactor decode-transaction to decode-transaction-message
